% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204183.R
\name{email}
\alias{email}
\title{A email dataset}
\description{
A email dataset of Hillary Clinton used to calculate the ItemRank.
}
\examples{
\dontrun{
data(email)
attach(email)
ir = itemrank(email, d = 0.85)
}
}
