% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204183.R
\name{itemrank}
\alias{itemrank}
\title{ItemRank algorithm using R}
\usage{
itemrank(pages, d = 1)
}
\arguments{
\item{pages}{Matrix of a graph, including the nodes with out-of-degree, nodes with in-of-degree and weight of edges.}

\item{d}{Damping factor, a positive value in (0,1), default 1.}
}
\value{
The ItemRank of the nodes in graph.
}
\description{
ItemRank algorithm using R
}
\examples{
\dontrun{
  weightdata <- t(matrix(c(1,2,0.3,
  1,3,0.7,
  2,3,1), nrow = 3))
  r2.1 = itemrank(weightdata, d = 1)
  r2.2 = itemrank(weightdata, d = 0.85)
}
}
