% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204183.R
\name{pagerank}
\alias{pagerank}
\title{PageRank algorithm using R}
\usage{
pagerank(pages, d = 1)
}
\arguments{
\item{pages}{Matrix of a graph, including the nodes with out-of-degree, nodes with in-of-degree and weight of edges.}

\item{d}{Damping factor, a positive value in (0,1), default 1.}
}
\value{
The PageRank of the nodes in graph.
}
\description{
PageRank algorithm using R
}
\examples{
\dontrun{
  pagedata <- t(matrix(c(1,2,1,
  1,3,1,
  2,3,1
  ), nrow = 3))
  r1.1 = pagerank(pagedata, d = 1)
  r1.2 = pagerank(pagedata, d = 0.85)
}
}
