% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204183.R
\name{personpagerank}
\alias{personpagerank}
\title{Personalized PageRank algorithm using R}
\usage{
personpagerank(pages, d = 1, root, is_using_out = FALSE)
}
\arguments{
\item{pages}{Matrix of a graph, including the nodes with out-of-degree, nodes with in-of-degree and weight of edges.}

\item{d}{Damping factor, a positive value in (0,1), default 1.}

\item{root}{The Personalized node with out-of-degree.}

\item{is_using_out}{Whether to use average out-of-degree as weight, default FALSE as not use.}
}
\value{
The Personalized PageRank of the root node to nodes with in-of-degree in graph.
}
\description{
Personalized PageRank algorithm using R
}
\examples{
\dontrun{
  persondata <- t(matrix(c(1,5,1,
  1,6,0.1,
  2,5,0.2,
  3,6,0.4,
  4,5,0.7), nrow = 3))
  r3.1 = personpagerank(persondata, d = 0.85, root = 1, is_using_out=TRUE)
  r3.2 = personpagerank(persondata, d = 0.85, root = 1, is_using_out=FALSE)
}
}
