% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PromC}
\alias{PromC}
\title{A probability matrix calculator}
\usage{
PromC(adjM, d)
}
\arguments{
\item{adjM}{the adjacency matrix}

\item{d}{the damping factor}
}
\value{
a probability matrix
}
\description{
A probability matrix calculator
}
\examples{
\dontrun{
A = matrix(c(0,1,0,1,0,0,0,1,1), ncol = 3)
P = PromC(A, 0.85)
}
}
