\name{SA23204183-package}
\alias{SA23204183-package}
\alias{SA23204183}
\docType{package}
\title{
  The Page Rank Family Function
}
\description{
  Calculates the pagerank values of nodes in the graph.
}
\details{
  PageRank is an algorithm proposed by Google for computing the ranking of web page importance. This package includes one function for generating graphs and two functions related to PageRank. However, PageRank and its derivative methods encompass a much broader range of algorithms.
}
\author{
Hanlu Zhang

Maintainer: Hanlu Zhang <zhanghanlu@mail.ustc.edu.cn>
}
\references{
  Page, L., Brin, S., Motwani, R., & Winograd, T. (1999). The PageRank citation ranking: Bringing order to the Web. Technical Report, Stanford InfoLab.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
